Module1/requirements.txt
requests>=2.31.0
urllib3>=1.26.0
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0